Replace files property with config property (#123)
Previously, you could only get/set the state of a playground with the files property. However, this means there's no way to directly set or access the importMap config, or other settings we may add in the future. Now there's a config property instead, which has exactly the same type as the JSON config.