Refactor playground worker API to be streaming (#163)
Defines a streaming API for composing builders into a pipeline within the web worker. Each builder receives and returns an async iterator of results. Results are files or diagnostics.

We now serve each asset as soon as it is ready, instead of waiting for the entire build to finish before loading. Most notably this means the index.html can start rendering concurrently with the TypeScript compile. The GET request for each asset blocks until it emerges from the stream (and 404s if the stream ends without that asset being found).