Add a WTR plugin for simulating CDNs like unpkg.com (#169)
This will be useful for testing an upcoming PR that proxies unpkg.com so that we can control the module specifiers of dependencies (#104).

It's a plugin for @web/test-runner that simulates an NPM CDN like unpkg.com. It extends the built-in @web/test-runner server, and can be re-configured from the test suite itself between each test using a server command (https://modern-web.dev/docs/test-runner/commands/).