Add util module with path, specifier, etc. utility functions (#170)
Adds some new utility functions, and moves some existing ones, into a new util.ts module, and adds unit tests for all of them.

These are mostly related to URL paths and module specifiers, but there's also MergedAsyncIterables and charToLineAndChar. This and some of the other currently unused ones will be used in upcoming PRs that are part of #104.