Add CachingCdn class (#174)
More progress towards improved playground dependency handling.

This PR adds a CachingCdn class, which is an interface to unpkg.com or similar CDNs. The main interesting thing it helps with is specifier canonicalization -- e.g. for resolving foo@^1.0.0 to foo@1.2.3/index.js, and it is intelligent about caching version ranges that have been already canonicalized (even for different paths).