Set Origin-Agent-Cluster header in SW (#179)
Chrome (and Firefox behind a flag) is able to allocate a separate process or thread to iframes, but only if the iframe and the parent are on separate sites, where a site is defined by protocol + top-level domain (note this does not include subdomains). However, there is a new header called Origin-Agent-Cluster, which allows a server to switch the key for process isolation from site to origin. This PR adds that header to service worker responses, and documents this issue for people who choose to change the default sandbox origin from unpkg to their own.