Force service worker updates on version mismatch (#178)
Previously, it was possible for the versions of a playground project element and its running service worker to be out of sync. This is because service workers are long-running, and only update themselves opportunistically. This could cause playgrounds to not work, or to behave strangely, for some time after an update.

We now include a version number check in the handshake between the playground project and the service worker. If the versions don't match, then we call ServiceWorkerRegistration.update to force an immediate update, and we wait for that update before we try and interact further with the service worker.