Improved types fetcher (#193)
A big refactor of the TypesFetcher class.

We now correctly check types when there are multiple versions of the same package simultaneously. Previously, if two versions of the same package appeared in the transitive import graph of a project, only some arbitrary version would be included. That was because we presented a completely flat virtual node_modules/ directory to the TypeScript compiler. We now support nested node_modules/ directories when there are version conflicts, much like a local NPM install. Fixes Allow multiple versions of the same typings to co-exist #192

We now respect the project's package.json if it exists, to determine which version of a package an import statement should be for. Previously, we ignored the package.json, so only import maps could be used to change the version from latest. Fixes Respect version constraints in a project package.json file when resolving types #184.