Creating a file that exists but is hidden now unhides the file. (#197)
If a file is marked hidden, then before this PR there was no way for the user to view or edit it. Creating a file with the same name was not allowed.

After this PR, then a user can reveal an existing hidden file by creating a file with the same name.

A little weird perhaps, but I think it's a decent solution. The motivating use case is that on lit.dev, we have a number of examples that have a package.json file pinning the project to the latest versions of Lit. However, we mark that package.json hidden so that it doesn't clutter up the list of files (especially when displayed as an inline sample). However, if a user wants to add a new dependency with a version constraint, or modify one of the existing Lit dependencies, there is currently no way to do it. Now, they'll be able to "create" a package.json, and then be able to see and modify the existing package.json.