Support package exports, and enable development condition (#200)
Adds support for Node-style package exports, with the "module", "import", and "development" conditions enabled.

This means, for example, that Lit will now run in development mode, and it is now possible to import Lion components using their recommended specifiers like @lion/button/define -- which requires the use of package exports.

Since the package exports resolution algorithm is a little complicated, and not actually fully documented, I've directly imported the module resolution functions from Node itself and am calling into that. It's checked into this repo (with a separate license header), rather than imported as a dependency, because 1. Node itself isn't available as an NPM package AFAIK, and 2. it contains a lot of non-web compatible code (require style imports, filesystem access, etc.), so a bunch of changes were needed.

Fixes #188