Fix positioning of no-service-worker error overlay (#231)
Fix an oversight in #224.

When service workers are not available, such as in Firefox in private browsing mode, we display a message over the preview. The overlay is absolutely positioned, and was intended to cover up just the preview (since the editor still works fine). However, since the preview is not necessarily itself positioned, the overlay was instead covering the next nearest positioned ancestor, which in the case of lit.dev was the entire window.

I didn't notice this because I only tested it in playground-ide, which positions the preview, but not when preview is used in a custom layout, such as in lit.dev.