fix: use lowercase content-type header value (#279)
Deployed to a server where .ts was being returned as video/MP2T instead of video/mp2t as required here. This particular header value is not case sensitive.

The type, subtype, and parameter names are not case sensitive. For example, TEXT, Text, and TeXt are all equivalent. Parameter values are normally case sensitive, but certain parameters are interpreted to be case- insensitive, depending on the intended use. (For example, multipart boundaries are case-sensitive, but the "access- type" for message/External-body is not case-sensitive.)

https://www.w3.org/Protocols/rfc1341/4_Content-Type.html