Add browser to export conditions and release 0.16.0 (#313)
Adds `browser` to the list of export conditions we use for resolving modules. This condition is listed in the "community conditions definitions" in the Node docs at https://nodejs.org/api/packages.html#community-conditions-definitions. We encountered it in a package we are using on lit.dev: https://unpkg.com/@floating-ui/dom@0.5.0/package.json