fix(cdn): change default value of baseUrl to empty string (#423)
`cdnBaseUrl` is settable via the project.json or via attribute/property. The property takes precedence over the project.json, and if neither is defined it defaults to unpkg.com. By having a default value for the property, then in order to take a value from the config, then the user would need to explicitly set the property / attribute to `''`.

Confirmed this works by setting the cdnBaseUrl in project/project.json in the configurator